Update model to use doubles instead of floats
Enable search from fab when search is visible

+++ app/src/main/java/com/hartcode/hartweather/data/record/WeatherRecord.java
@@ -11,17 +11,17 @@ public class WeatherRecord extends SugarRecord {
     @Unique
     public long id;
     public int cityId;
-    public float lat;
-    public float lon;
+    public double lat;
+    public double lon;
     public String cityName;
     public String main;
     public String description;
     public String icon;
-    public float temp;
-    public float pressure;
-    public float humidity;
-    public float temp_min;
-    public float temp_max;
+    public double temp;
+    public double pressure;
+    public double humidity;
+    public double temp_min;
+    public double temp_max;
     public long lastUpdate;
 
     public WeatherRecord()
@@ -29,7 +29,7 @@ public class WeatherRecord extends SugarRecord {
 
     }
 
-    public WeatherRecord(long id, int cityId, float lat, float lon, String cityName, String main, String description, String icon, float temp, float pressure, float humidity, float temp_min, float temp_max, long lastUpdate)
+    public WeatherRecord(long id, int cityId, double lat, double lon, String cityName, String main, String description, String icon, double temp, double pressure, double humidity, double temp_min, double temp_max, long lastUpdate)
     {
         this.id = id;
         this.cityId = cityId;
